+++ .dockerignore
@@ -22,7 +22,6 @@
 
 # Logs
 **/logs
-!**/src/logs
 **/*.log
 **/npm-debug.log*
 **/yarn-debug.log*
