+++ .gitignore
@@ -5,7 +5,6 @@
 
 # Logs
 **/logs
-!**/src/logs
 **/*.log
 **/npm-debug.log*
 **/yarn-debug.log*
