+++ libs/@local/telemetry/src/logging/console.rs
