+++ libs/@local/telemetry/src/logging/formatter.rs
